requirements.txt
requests~=2.30.0
python-dotenv~=1.0.0
bs4~=0.0.1
beautifulsoup4~=4.12.2